DOM: Reverse order of nodes in coalesce
Overriden value should be evaluated first, otherwise override will never work.
Move coalesce to appropriate merge*.go files and add unit test for this situation
